tests/wpt/web-platform-tests/resource-timing/no-entries-for-cross-origin-css-fetched-memory-cache.sub.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Make sure that resources fetched by cross origin CSS are not in the timeline.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<body>
<script>
// This test is mostly the same as no-entries-for-cross-origin-css-fetched.html
// but loads the subresources from another Document (<iframe>) earlier before
// they are loaded via cross-origin CSSs, because the ResourceTiming processing
// goes through a different code path in Chrome when served from MemoryCache
// (`ResourceFetcher::DidLoadResourceFromMemoryCache`).
function runTest() {
    const link = document.createElement("LINK");
    link.rel = "stylesheet";
    link.id = "cross_origin_style";
    link.href = "http://{{hosts[][www1]}}:{{ports[http][1]}}/resource-timing/resources/nested.css";
    document.body.appendChild(link);
}
</script>
<script>
  const t = async_test("Make sure that resources fetched by cross origin CSS are not in the timeline.");
  window.addEventListener("load", function() {
    // A timeout is needed as entries are not guaranteed to be in the timeline before onload triggers.
    t.step_timeout(function() {
          const url = (new URL(document.getElementById("cross_origin_style").href));
          const prefix = url.protocol + "//" + url.host;
          assert_equals(performance.getEntriesByName(prefix + "/resource-timing/resources/resource_timing_test0.css?id=n1").length, 0, "Import should not be in timeline");
          assert_equals(performance.getEntriesByName(prefix + "/fonts/Ahem.ttf?id=n1").length, 0, "Font should not be in timeline");
          assert_equals(performance.getEntriesByName(prefix + "/resource-timing/resources/blue.png?id=n1").length, 0, "Image should not be in timeline");
          t.done();
    }, 200);
  });
</script>
<iframe src="resources/no-entries-for-cross-origin-css-fetched-memory-cache-iframe.sub.html" onload="runTest()"></iframe>
<ol>Some content</ol>
</body>
